Add files via upload
Here, we investigate a novel, general, a machine learning method for deducing attribute-specific perception by mining the brand’s social tweets onTwitter. The method provides a reliable, and scalable method for the detection of brand perceptions, and estimates a subtle scope for future advances in understanding relationships between customers and brands.

We aim to compute brand perception across all states, and for that, we have used logistic regression and to find its performance measurement we have applied the ROC-AUC curve. This classifier has a performance accuracy of 79%.